+++ three/renderers/impl/gl_renderer.cpp
@@ -902,7 +902,7 @@ void GLRenderer::initDirectBuffers( Geometry& geometry ) {
 
     auto& attribute = a.second;
     attribute.buffer = _gl.CreateBuffer();
-    attribute.numItems = attribute.array.size();
+    attribute.numItems = (int)attribute.array.size();
 
     _gl.BindAndBuffer( type, attribute.buffer, attribute.array, GL_STATIC_DRAW );
 
@@ -2751,7 +2751,7 @@ void GLRenderer::render( Scene& scene, Camera& camera, const GLRenderTarget::Ptr
 
   auto& renderList = scene.__glObjects;
 
-  for ( size_t i = 0; i < renderList.size(); i++) {
+  for ( int i = 0; i < renderList.size(); i++) {
 
     auto& glObject =  renderList[i];
     auto& object = *glObject.object;
